refactor: presigned URL 기반 파일 다운로드 로직 공통 유틸로 분리 (#356)
